Update patch set 3
Patch Set 3:

Steve,

Forgive my new-ness, but are the samsung msm8960 commits something that might also benefit the htc One family as well (htc msm8960), or are the changes vendor-specific?

I've been keeping a close eye on the S3 (intl) development vs. the HTC Ville, and wondering if there are common traits to improve both...

Patch-set: 3