Update patch set 1
Patch Set 1: No score

I believe the whole reason it was split off was so that more then the 8960 would use the common qcom (such as the 8660)

seeing how the only value currently used is the PLATFORM_BOARD, i could see just defining that in msm-common instead of including the entire boardconfig.

My intent was to resolve building of what was already committed to git last night.

Patch-set: 1