Update patch set 2
Patch Set 2:

I don't believe it's needed.  The only reason to inherit the BoardConfig currently is to pass the TARGET_BOARD_PLATFORM value to the qcom-common.mk so it knows which libs to build.

setting just TARGET_BOARD_PLATFORM := msm8960 would work as well but if further values are needed later, its one less change to just leave it pointing to the board config.

Ultimately I can't say what the best method would be.  The qcom-common was merged in early this morning and I noticed this when trying to build today and ended up w/out any of the msm8960 libs

Patch-set: 2