Update patch set 2
Patch Set 2:

If you turn on that settings area, it controls bluetooth docks only.

It's currently disabled in d2-common:
https://github.com/CyanogenMod/android_device_samsung_d2-common/blob/jellybean/overlay/packages/apps/Settings/res/values/config.xml

In CM9, the checkbox for dock USB audio routing was built into that menu.  in CM10, Steven Harper moved it to DeviceSettings.. so this is porting that work to DeviceSettings for d2.

Patch-set: 2