Update patch set 1
Patch Set 1:

Hmm, so the property subverts the framework audio path API as it's a non-standard extension.  That's why I kept the implemenation in a device repo.  I'm not sure how it's done on other devices.

If nothing else, we'll have to add "persist.audio." to system/core/property_services.c which arcee doesn't like so much.

If you prefer that we drop the current implementation into Phone with a "config_has..." option, I can do that.  Otherwise we'd have to think of the "right" way to extend the framework API, if we want to go that route.

Patch-set: 1