Update patch set 5
Patch Set 5: I would prefer that you didn't submit this

I actually just tested.  I believe the if for the dock settings needs to be moved up as a else if before the return, around line 67 of soundactivityfragment.  I would line-comment, but can't seem to figure out how to from my phone.  If no one gets to it, I'll do it when I'm home in a couple hours.

Patch-set: 5