Update patch set 1
Patch Set 1:

Unfortunately the advanced settings hook is already being used by the SGS III apk. There is no way to disinherit it. 

The settings apk already takes into account the existence of different downstream devices, for example in TouchscreenSensitivity.java, a device which does not have a Melfas touchscreen will not see the option. In SensorsFragmentActivity, a device which does not have BLN support will not see the option. This patch uses the current code behavior and tests whether a device is compatible before displaying the option. Users without a hard keyboard will never even be aware the option exists.

Patch-set: 1