Update patch set 1
Patch Set 1: Code-Review-2

Bleh, sorry for adding you all to the apex commit (if the email got generated).

Ok, here we go, try 2, now that we've moved it to msm8960-common for the jf devices =)

I moved media_profiles.xml to apexqtmo and expressatt device trees, to override d2, didn't seem right to leave a camera/ folder in there with just media_profiles.xml in it.

Need the ifdef here tested for apex and express, you guys don't want ZSL.

There are a couple little bugs we need to work out for d2. I have also not tested this on jf yet, will tonight.

Requirements:

http://review.cyanogenmod.org/#/c/55269/

http://review.cyanogenmod.org/#/c/55270/

http://review.cyanogenmod.org/#/c/55271/

http://review.cyanogenmod.org/#/c/55272/

http://review.cyanogenmod.org/#/c/55273/

https://github.com/invisiblek/proprietary_vendor_samsung/commit/27d419280e89d8f5c02a81513127d94bb359fa91

Patch-set: 1