Update patch set 4
Patch Set 4:

jt1134: I've personally verified Snapchat doesn't work on apex and d2 in video mode with camcorder mode on. Hangouts is also said to be affected. Right now it's off, which breaks the built-in app on d2 (doesn't seem to affect apex). 

I'm hoping the fix to camera2 gets merged but in the meantime it's interesting to see if I can fix it in the wrapper.

Patch-set: 4