Update patch set 2
Patch Set 2:

Right now this is off altogether in the wrapper, it causes ffc distortions in some apps, ie Snapchat although Camera2 won't function without it (which is why we had to add it to Camera). Jf doesn't seem to need it at all.

Enabling it in the wrapper for rear cam has no benefits either way that I can tell although I have yet to test camera2 without it turned on locally. If we can reduce reliance on Camera2 hacks it's better in the longrun.

Patch-set: 2