msm8960: fix samsung camcorder mode the right way
A lot of these modes are inter-dependent. Don't mess with Samsung camera
mode if it's already been set.

There is also a problem with apps which use the FFC for video chat, that
don't set the recording hint properly. Work around these by guessing the
mode from the camera internal zsl mode flag.

Change-Id: I8f1fc611a2f100d0438447c25103fc44917c5d87