Update patch set 1
Patch Set 1:

Don't worry, jt1134, this is for d2 only. apex/express should be working for everything but snapchat and instagram. I don't really have any solid leads on them, I think snapchat is doing it to itself (it tries to set the modes but apex doesn't support them)

Patch-set: 1