Update patch set 2
Patch Set 2:

Thanks for the reply.  I assume you mean the newer stock rom on d2lte...?  Does the newer firmware or blobs require the change?

This repo is also used for apexqtmo, which I have recently inherited.  The mount change does seem to work, but I also have a pending change to separate the modem firmware links because the apexqtmo modem files differ from d2lte.  So in the future a change like this will need to be coordinated and tested across multiple device repos.

Patch-set: 2