Update patch set 17
Patch Set 17:

I only have one TARGET_SPECIFIC_HEADER_PATH defined, so it can't be that. However, I just found that there's a bug in the parameter override patch. It's supposed to be #include "camera/CameraParametersExtra.h" and not #include <camera/CameraParametersExtra.h>

After changing that, it compiles fine.

Patch-set: 17