Update patch set 8
Patch Set 8: -Code-Review

Oh ok is that why it takes a while for the radio to actually kick in once it's booted 

I've been using your commits for a LiquidSmooth build and I'm on a build right now that's working pretty well only this I've notice is radio takes some time to kick in,, and for some reason 3g and WiFi are turning on at the same time not to sure what's causing that. Also I haven't been able to get the camera to work yet either but I maybe missing a commit somewhere. 

Anything I can do to help you out just let me know

Patch-set: 8