Update patch set 15
Patch Set 15:

Ok, other than using macros for the permissions on much of this, it should be pretty much ready to go.

Please test with this and the next commit to re-enable enforcing.

I imagine express/apex are going to need more rules added. D2 should be ready to ship though

Patch-set: 15