msm8960-common: let the h@x continue
Checks a parameters string for validity by removing everything that
isn't an alpha-numeric character and ending it with a null termination.
THANKS SAMSUNG.
Code gratuitously nicked from
http://rosettacode.org/wiki/Strip_control_codes_and_extended_characters_from_a_string#C
The wiki is licensed GFDL 1.2, which *is* a free license but it isn't
exactly portable, there's like 3 pages of disclaimers that are supposed
to accompany any copy of a work, etc.
So, if you want to see the disclaimers, click the link and read them.

Re-add debug logging. I didn't add logging for the fixed set of
get_parameters, because we're directly returning the parameters object
instead of copying it to a new string first like we did in the past. I
guess I could use params.dump, but I hate the format it comes out in the
logs. Furthermore, I *know* what I'm doing to the get_parameters so I
don't need to look at it.

Fix potential memory leak in get_parameters

Consolidate all d2 specific code under a single define

Change-Id: I01670daab9fc1665e61db58aa98900072824f75e