msm8960-common: let the h@x continue
Say good bye to Samsung Camera Mode. This incredibly ugly hack that had
to differentiate between different *apps* using the camera is no longer
necessary.

Say hello to strip(), a function that checks a parameters string for
validity by removing everything that isn't an alpha-numeric character
and ending it with a null termination. THANKS SAMSUNG.
Code gratuitously nicked from
http://rosettacode.org/wiki/Strip_control_codes_and_extended_characters_from_a_string#C
The wiki is licensed GFDL 1.2, which *is* a free license but it isn't
exactly portable, there's like 3 pages of disclaimers that are supposed
to accompany any copy of a work, etc.
So, if you want to see the disclaimers, click the link and read them.

Re-add debug logging. I didn't add logging for the fixed set of
get_parameters, because we're directly returning the parameters object
instead of copying it to a new string first like we did in the past. I
guess I could use params.dump, but I hate the format it comes out in the
logs. Furthermore, I *know* what I'm doing to the get_parameters so I
don't need to look at it.

Re-add a local put_parameters for the get_parameters. We're copying the
parameters string into a CameraParameters object called params, I
*think* this is a new pointer, I will have to check before this is
merged. Otherwise I'm creating a use after free.

Add Magic 1508 for ZSL mode. Might remove before merging.

Change-Id: I01670daab9fc1665e61db58aa98900072824f75e