Update patch set 1
Patch Set 1:

hammerhead also uses bcmdhd and removed the interface. But they're inheriting stock wpa_supplicant (which I can't be arsed to look for in the tree but I suspect also contains the change mako's does. You should probably look there and confirm and if it does make that change instead of reverting this)

Patch-set: 1