Update patch set 1
Patch Set 1:

There is a time limit for optimization of apk. Some apps like play services takes a lot of time, so it doesn't gets optimized properly due to time limit. So it tries to optimize on each both.It also happens when it's updated. By disabling watchdog we let the apk to get optimized completely.

Patch-set: 1